Homework_11/requirements.txt
requests==2.31.0
bs4==4.11.1
numpy==1.21.5
pandas==1.4.4
matplotlib==3.5.2
seaborn==0.11.2
python-dotenv==1.0.0
